refactor: move mongo logic to config
